rename IntegrationExplorer to ExplorerLinkBuilder, fold getExplorerURL in
ExplorerLinkBuilder is now the single home for every URL the app shows
"View on Explorer" / "View Progress" with — including the chain
explorer URL formerly at Endpoint.getExplorerURL. Endpoint keeps only
its actual API/RPC endpoints; the integration-tracker concerns (LI.FI,
Maya, THORChain, plus the chain-explorer fallback) all live in one
place.

- Move getExplorerURL(chain:txid:) from Endpoint to ExplorerLinkBuilder
  unchanged.
- Rename IntegrationExplorer -> ExplorerLinkBuilder (file + type).
- Rename IntegrationExplorerTests -> ExplorerLinkBuilderTests.
- Update 19 call sites of Endpoint.getExplorerURL and 5 call sites of
  IntegrationExplorer.

Build green; lint clean on changed files.
