+++ VultisigApp/VultisigApp/Core/Utils/Endpoint.swift
@@ -475,94 +475,6 @@ class Endpoint {
     static let tronEvmServiceRpc = "https://api.vultisig.com/tron-rpc"
     static let tronWalletApi = "https://api.vultisig.com/tron"
 
-    static func getExplorerURL(chain: Chain, txid: String) -> String {
-        switch chain {
-        case .bitcoin:
-            return "https://mempool.space/tx/\(txid)"
-        case .bitcoinCash:
-            return "https://blockchair.com/bitcoin-cash/transaction/\(txid)"
-        case .litecoin:
-            return "https://blockchair.com/litecoin/transaction/\(txid)"
-        case .dogecoin:
-            return "https://blockchair.com/dogecoin/transaction/\(txid)"
-        case .dash:
-            return "https://blockchair.com/dash/transaction/\(txid)"
-        case .zcash:
-            return "https://blockchair.com/zcash/transaction/\(txid)"
-        case .thorChain:
-            return "https://runescan.io/tx/\(txid.stripHexPrefix())"
-        case .thorChainChainnet:
-            return "https://runescan.io/tx/\(txid.stripHexPrefix())?network=chainnet"
-        case .thorChainStagenet:
-            return "https://runescan.io/tx/\(txid.stripHexPrefix())?network=stagenet"
-        case .solana:
-            return "https://orb.helius.dev/tx/\(txid)"
-        case .ethereum:
-            return "https://etherscan.io/tx/\(txid)"
-        case .gaiaChain:
-            return "https://www.mintscan.io/cosmos/tx/\(txid)"
-        case .dydx:
-            return "https://www.mintscan.io/dydx/tx/\(txid)"
-        case .kujira:
-            return "https://finder.kujira.network/kaiyo-1/tx/\(txid)"
-        case .avalanche:
-            return "https://snowtrace.io/tx/\(txid)"
-        case .bscChain:
-            return "https://bscscan.com/tx/\(txid)"
-        case .mayaChain:
-            return "https://www.explorer.mayachain.info/tx/\(txid)"
-        case .arbitrum:
-            return "https://arbiscan.io/tx/\(txid)"
-        case .base:
-            return "https://basescan.org/tx/\(txid)"
-        case .optimism:
-            return "https://optimistic.etherscan.io/tx/\(txid)"
-        case .polygon, .polygonV2:
-            return "https://polygonscan.com/tx/\(txid)"
-        case .blast:
-            return "https://blastscan.io/tx/\(txid)"
-        case .cronosChain:
-            return "https://cronoscan.com/tx/\(txid)"
-        case .sui:
-            return "https://suiscan.xyz/mainnet/tx/\(txid)"
-        case .polkadot:
-            return "https://assethub-polkadot.subscan.io/extrinsic/\(txid)"
-        case .bittensor:
-            let hash = txid.hasPrefix("0x") ? txid : "0x\(txid)"
-            return "https://taostats.io/extrinsic/\(hash)"
-        case .zksync:
-            return "https://explorer.zksync.io/tx/\(txid)"
-        case .ton:
-            return "https://tonviewer.com/transaction/\(txid)"
-        case .osmosis:
-            return "https://www.mintscan.io/osmosis/tx/\(txid)"
-        case .terra:
-            return "https://www.mintscan.io/terra/tx/\(txid)"
-        case .terraClassic:
-            return "https://finder.terra.money/classic/tx/\(txid)"
-        case .noble:
-            return "https://www.mintscan.io/noble/tx/\(txid)"
-        case .ripple:
-            return "https://xrpscan.com/tx/\(txid)"
-        case .akash:
-            return "https://www.mintscan.io/akash/tx/\(txid)"
-        case .tron:
-            return "https://tronscan.org/#/transaction/\(txid)"
-        case .ethereumSepolia:
-            return "https://sepolia.etherscan.io/tx/\(txid)"
-        case .cardano:
-            return "https://cardanoscan.io/transaction/\(txid)"
-        case .mantle:
-            return "https://explorer.mantle.xyz/tx/\(txid)"
-        case .hyperliquid:
-            return "https://liquidscan.io/tx/\(txid)"
-        case .sei:
-            return "https://seiscan.io/tx/\(txid)"
-        case .qbtc:
-            return ""
-        }
-    }
-
     static func getExplorerByAddressURL(chain: Chain, address: String) -> String? {
         switch chain {
         case .bitcoin:
