+++ VultisigApp/VultisigApp/Features/Referral/Views/ReferralTransactionOverviewView.swift
@@ -22,7 +22,7 @@ struct ReferralTransactionOverviewView: View {
                     amountCrypto: "\(sendTx.amount) \(sendTx.coin.ticker)",
                     amountFiat: sendTx.amountInFiat,
                     hash: hash,
-                    explorerLink: Endpoint.getExplorerURL(chain: sendTx.coin.chain, txid: hash),
+                    explorerLink: ExplorerLinkBuilder.getExplorerURL(chain: sendTx.coin.chain, txid: hash),
                     memo: sendTx.memo,
                     isSend: false,
                     fromAddress: sendTx.fromAddress,
